Refactory Cryptografy_Service and Create Cryptography_Sample to Storage old code
